+++ src/main/kotlin/smalltalk/backend/ChatService.kt
@@ -2,6 +2,7 @@ package smalltalk.backend
 
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import org.springframework.stereotype.Service
+import smalltalk.backend.presentation.dto.chatmessage.ChatMessage
 
 
 @Service
