+++ src/main/kotlin/smalltalk/backend/presentation/controller/ChatRoomController.kt
@@ -1,8 +1,10 @@
-package smalltalk.backend
+package smalltalk.backend.presentation.controller
 
 import org.springframework.messaging.handler.annotation.DestinationVariable
 import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.stereotype.Controller
+import smalltalk.backend.presentation.dto.chatmessage.ChatMessage
+import smalltalk.backend.ChatService
 
 @Controller
 class ChatRoomController (
