+++ src/main/kotlin/smalltalk/backend/presentation/dto/chatmessage/ChatMessage.kt
@@ -1,4 +1,4 @@
-package smalltalk.backend
+package smalltalk.backend.presentation.dto.chatmessage
 
 
 data class ChatMessage (
